+++ pkg/resource/ack/ec2/security_group.go
@@ -13,21 +13,6 @@ type SecurityGroupOptions struct {
 }
 
 func NewSecurityGroupResource() *resource.Resource {
-	res := &resource.Resource{
-		Command: cmd.Command{
-			Name:        "security-group",
-			Description: "EC2 Security Group",
-			Aliases:     []string{"security-groups", "sg"},
-			CreateArgs:  []string{"NAME"},
-		},
-
-		Manager: &manifest.ResourceManager{
-			Template: &template.TextTemplate{
-				Template: securityGroupYamlTemplate,
-			},
-		},
-	}
-
 	options := &SecurityGroupOptions{
 		CommonOptions: resource.CommonOptions{
 			Name:          "ack-ec2-security-group",
@@ -36,21 +21,33 @@ func NewSecurityGroupResource() *resource.Resource {
 		},
 	}
 
-	flags := cmd.Flags{
-		&cmd.StringFlag{
-			CommandFlag: cmd.CommandFlag{
-				Name:        "desc",
-				Description: "description",
-				Required:    true,
-			},
-			Option: &options.Description,
+	return &resource.Resource{
+		Command: cmd.Command{
+			Name:        "security-group",
+			Description: "EC2 Security Group",
+			Aliases:     []string{"security-groups", "sg"},
+			CreateArgs:  []string{"NAME"},
 		},
+
+		CreateFlags: cmd.Flags{
+			&cmd.StringFlag{
+				CommandFlag: cmd.CommandFlag{
+					Name:        "desc",
+					Description: "description",
+					Required:    true,
+				},
+				Option: &options.Description,
+			},
+		},
+
+		Manager: &manifest.ResourceManager{
+			Template: &template.TextTemplate{
+				Template: securityGroupYamlTemplate,
+			},
+		},
+
+		Options: options,
 	}
-
-	res.Options = options
-	res.CreateFlags = flags
-
-	return res
 }
 
 const securityGroupYamlTemplate = `---
